change int to lon

+++ src/main/java/com/socialsend/sendapi/service/Transactions.java
@@ -355,7 +355,7 @@ public class Transactions {
 			return res;
 		}
 		
-		int id = db.insertPayment(param);
+		long id = db.insertPayment(param);
 		
 		if(id == 0) {
 			res = new Response<NewPaymentResponse>(null);
